Rename the tag SDK files
