+++ pkg/sdk/tags_gen.go
@@ -3,9 +3,6 @@ package sdk
 import (
 	"context"
 	"database/sql"
-	"fmt"
-	"slices"
-	"strings"
 	"time"
 )
 
@@ -56,30 +53,6 @@ type createTagOptions struct {
 
 type AllowedValues struct {
 	Values []StringAllowEmpty `ddl:"list,comma"`
-}
-
-type TagPropagation string
-
-const (
-	TagPropagationNone                        TagPropagation = "NONE"
-	TagPropagationOnDependency                TagPropagation = "ON_DEPENDENCY"
-	TagPropagationOnDataMovement              TagPropagation = "ON_DATA_MOVEMENT"
-	TagPropagationOnDependencyAndDataMovement TagPropagation = "ON_DEPENDENCY_AND_DATA_MOVEMENT"
-)
-
-var AllTagPropagationValues = []TagPropagation{
-	TagPropagationNone,
-	TagPropagationOnDependency,
-	TagPropagationOnDataMovement,
-	TagPropagationOnDependencyAndDataMovement,
-}
-
-func ToTagPropagation(s string) (TagPropagation, error) {
-	tp := TagPropagation(strings.ToUpper(s))
-	if !slices.Contains(AllTagPropagationValues, tp) {
-		return "", fmt.Errorf("invalid tag propagation value: %s", tp)
-	}
-	return tp, nil
 }
 
 type TagPropagate struct {
